Estruturando a pagina, foi criado uma page Home e um componente HeaderMenu.
